Tomba (tomba v0.1.0)

Documentation for Tomba.

Tomba: Tomba.io is an Email Finder for B2B sales and email marketing This is the Lua client library for the Email Finder API. It allows you to lookup the following details :

Domain Search (Search emails are based on the website You give one domain name and it returns all the email addresses found on the internet.)

Email Finder (This API endpoint generates or retrieves the most likely email address from a domain name, a first name and a last name..)

Email Verifier (checks the deliverability of a given email address, verifies if it has been found in our database, and returns their sources.)

Email Sources (Find email address source somewhere on the web .)

Company Domain autocomplete (Company Autocomplete is an API that lets you auto-complete company names and retrieve logo and domain information.)

Types

http_opts()

Specs

http_opts() :: [
  body: binary() | charlist(),
  headers: [{atom() | String.Chars.t(), String.Chars.t()}],
  query: %{optional(String.Chars.t()) => String.Chars.t()},
  timeout: timeout(),
  basic_auth: {List.Chars.t(), List.Chars.t()},
  stream_to: pid() | port() | atom() | {atom(), node()},
  direct: pid() | port() | atom() | {atom(), node()},
  ibrowse: keyword(),
  auto_sni: boolean(),
  follow_redirects: boolean()
]

List of options to all request functions.

body - request body

headers - HTTP headers (e.g. ["Accept": "application/json"])

query - URL query string (e.g. %{page: 1})

timeout - timeout in milliseconds

basic_auth - basic auth credentials (e.g. {"user", "password"})

stream_to - if you want to make an async request, reference to the process

direct - if you want to use ibrowse's direct feature, reference to
       the worker spawned by `spawn_worker_process/2` or `spawn_link_worker_process/2`

ibrowse - options for ibrowse

auto_sni - if true and the URL is https, configures the server_name_indication ibrowse/ssl option
         to be the host part of the requestedURL

follow_redirects - if true and a response is a redirect, re-requests with header[:Location]

http_result()

Specs

http_result() ::
  HTTPotion.Response.t()
  | %HTTPotion.AsyncResponse{id: term()}
  | %HTTPotion.ErrorResponse{message: term()}

Result returned from request/3, get/2, post/2, put/2, etc.

http_result_bang()

Specs

http_result_bang() ::
  HTTPotion.Response.t() | %HTTPotion.AsyncResponse{id: term()}

Result returned from request!/3, get!/2, post!/2, put!/2, etc.

Functions

account(this)

Returns information about the current account.

Parameters

this - Tomba

## Examples

{:ok, result} = Tomba.account(Tomba)

account_path()

autocomplete(this, query)

Company Autocomplete is an API that lets you auto-complete company names and retrieve logo and domain information.

Parameters

this - Tomba

query - A string name company or website.

## Examples

{:ok, result} = Tomba.autocomplete(Tomba,"google")

autocomplete_path()

call(this, path, params)

call to server

Parameters

this - Tomba

path - A string specific path.

params- query params

count(this, domain)

Returns total email addresses we have for one domain.

Parameters

this - Tomba

domain - A string specific domain.

## Examples

{:ok, result} = Tomba.count(Tomba,"tomba.io")

count_path()

default_base_url()

delete(url, options \\ [])

Specs

delete(String.Chars.t(), http_opts()) :: http_result()

A shortcut for request(:delete, url, options).

delete!(url, options \\ [])

Specs

delete!(String.Chars.t(), http_opts()) :: http_result_bang()

A shortcut for request!(:delete, url, options).

domain_search(this, domain)

Search emails are based on the website You give one domain name and it returns all the email addresses found on the internet.

Parameters

this - Tomba

domain - A string specific domain.

## Examples

{:ok, result} = Tomba.domain_search(Tomba,"tomba.io")

email_finder(this, domain, fname, lname)

Generates or retrieves the most likely email address from a domain name, a first name and a last name.

Parameters

this - Tomba

domain - A string domain name of the company, used for emails. For example, "tomba.com".

fname - A string The person's first name. It doesn't need to be in lowercase.

lname - A string The person's last name. It doesn't need to be in lowercase.

## Examples

{:ok, result} = Tomba.email_finder(Tomba,"stripe.com", "fname", "lname")

email_sources(this, email)

Find email address source somewhere on the web.

Parameters

this - Tomba

email - A string email address you want to find sources.

## Examples

{:ok, result} = Tomba.email_sources(Tomba,"b.mohamed@tomba.io")

email_verifier(this, email)

Verify the deliverability of an email address.

Parameters

this - Tomba

email - A string email address you want to verify.

## Examples

{:ok, result} = Tomba.email_verifier(Tomba,"b.mohamed@tomba.io")

finder_path()

get(url, options \\ [])

Specs

get(String.Chars.t(), http_opts()) :: http_result()

A shortcut for request(:get, url, options).

get!(url, options \\ [])

Specs

get!(String.Chars.t(), http_opts()) :: http_result_bang()

A shortcut for request!(:get, url, options).

handle_response(response)

head(url, options \\ [])

Specs

head(String.Chars.t(), http_opts()) :: http_result()

A shortcut for request(:head, url, options).

head!(url, options \\ [])

Specs

head!(String.Chars.t(), http_opts()) :: http_result_bang()

A shortcut for request!(:head, url, options).

is_redirect(response)

logs(this)

Returns a your last 1,000 requests you made during the last 3 months.

Parameters

this - Tomba

## Examples

{:ok, result} = Tomba.logs(Tomba)

logs_path()

new(key, secret)

Tomba Constructor

Parameters

key - Tomba api key.

secret - Tomba secret key.

options(url, options \\ [])

Specs

options(String.Chars.t(), http_opts()) :: http_result()

A shortcut for request(:options, url, options).

options!(url, options \\ [])

Specs

options!(String.Chars.t(), http_opts()) :: http_result_bang()

A shortcut for request!(:options, url, options).

patch(url, options \\ [])

Specs

patch(String.Chars.t(), http_opts()) :: http_result()

A shortcut for request(:patch, url, options).

patch!(url, options \\ [])

Specs

patch!(String.Chars.t(), http_opts()) :: http_result_bang()

A shortcut for request!(:patch, url, options).

post(url, options \\ [])

Specs

post(String.Chars.t(), http_opts()) :: http_result()

A shortcut for request(:post, url, options).

post!(url, options \\ [])

Specs

post!(String.Chars.t(), http_opts()) :: http_result_bang()

A shortcut for request!(:post, url, options).

process_options(options)

process_request_body(body)

process_request_headers(headers)

process_request_headers(headers, body)

process_request_headers(headers, body, options)

process_response_body(body)

process_response_body(headers, body)

process_response_chunk(body)

process_response_headers(headers)

process_response_location(response)

process_status_code(status_code)

process_url(url)

process_url(url, options)

put(url, options \\ [])

Specs

put(String.Chars.t(), http_opts()) :: http_result()

A shortcut for request(:put, url, options).

put!(url, options \\ [])

Specs

put!(String.Chars.t(), http_opts()) :: http_result_bang()

A shortcut for request!(:put, url, options).

redirect_method(response, method)

request(method, url, options \\ [])

Specs

request(atom(), String.Chars.t(), http_opts()) :: http_result()

Sends an HTTP request.

See the type documentation of http_opts for a description of options.

request(method, url, body, headers, options)

This function is deprecated. Use request/3 instead.

request!(method, url, options \\ [])

Specs

request!(atom(), String.Chars.t(), http_opts()) :: http_result_bang()

Like request/3, but raises HTTPotion.HTTPError if failed.

request_direct(conn_pid, method, url, body \\ "", headers \\ [], options \\ [])

This function is deprecated. Use request/3 with 'direct' option instead.

response_ok(response)

search_path()

sources_path()

spawn_link_worker_process(url, options \\ [])

Starts a linked worker process for use with the direct option.

spawn_worker_process(url, options \\ [])

Starts a worker process for use with the direct option.

start()

Ensures that HTTPotion and its dependencies are started.

status(this, domain)

Returns domain status if is webmail or disposable.

Parameters

this - Tomba

domain - A string specific path.

## Examples

{:ok, result} = Tomba.status(Tomba,"gmail.com")

status_path()

stop_worker_process(pid)

Stops a worker process started with spawn_worker_process/2 or spawn_link_worker_process/2.

transformer(target, method, url, options)

usage(this)

Check your monthly requests.

Parameters

this - Tomba

## Examples

{:ok, result} = Tomba.usage(Tomba)

usage_path()

verifier_path()