Validation.Keyvalidation (tomba v0.1.0)

Tomba email finder elixir Client Library

Tomba: Tomba.io is an Email Finder for B2B sales and email marketing This is the Lua client library for the Email Finder API. It allows you to lookup the following details :

Domain Search (Search emails are based on the website You give one domain name and it returns all the email addresses found on the internet.)

Email Finder (This API endpoint generates or retrieves the most likely email address from a domain name, a first name and a last name..)

Email Verifier (checks the deliverability of a given email address, verifies if it has been found in our database, and returns their sources.)

Email Sources (Find email address source somewhere on the web .)

Company Domain autocomplete (Company Autocomplete is an API that lets you auto-complete company names and retrieve logo and domain information.)
Documentation for Key validation.

Functions

validate(str)
(since 0.1.0)

Helper method for validating tomba key and secret

## Parameters

str - Tomba api key or api secret.